requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
aiohttp==3.9.3
redis==5.0.1
prometheus-client==0.19.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9 